change build script

+++ build.sh
@@ -1,11 +1,23 @@
 #!/bin/bash
 shopt -s extglob
 
-deffile=${1-gfae.def}
+build_location=$1
+deffile=mp2rage_recon-all.def
 container_fname=$(basename $deffile .def).sif
 
+if [ ! -f license.txt ]
+then
+    echo "Freesurfer license file missing."
+    exit
+fi
+
 # BUILD CONTAINER
-tmpdir=$(mktemp -d)
+if [ -z "$build_location" ]
+then
+      tmpdir=$(mktemp -d)
+else
+      tmpdir=$(mktemp -d -p ${build_location})
+fi
 cwd=$(pwd)
 
 cp -r !(*.sif) $tmpdir/
